+++ changes.md
@@ -1,1 +1,2 @@
 * Add option to break only parts of the tree that are above the cut.
+* Add GUI to change configuration (Main menu > Mods > FallingTree > Config).